improved datasegment signature
